requirements.txt
langchain
duckduckgo-search
python-dotenv
schedule
slack-sdk
langchain-community
requests
schedule
langchain-google-genai
transformers
torch
accelerate
einops
xformers
newsapi-python